Typescriptify some files
